requirements.txt
numpy>=1.26.0
pandas>=2.1.0
yfinance>=0.2.36
requests>=2.31.0
beautifulsoup4>=4.12.2
schedule>=1.2.1
pyyaml>=6.0.1
scikit-learn>=1.3.2
matplotlib>=3.8.2
polygon-api-client>=1.13.3
jupyter>=1.0.0
notebook>=7.0.6
statsmodels>=0.14.1  # 添加这行